+++ realgui/engine/font/font_ttf.h
@@ -195,7 +195,7 @@ void gui_font_ttf_adapt_rect(gui_text_t *text, gui_text_rect_t *rect);
  * @return Character width.
  */
 uint32_t gui_get_ttf_char_width(void *content, void *font_bin_addr, TEXT_CHARSET charset,
-                                uint8_t font_height);
+                                uint16_t font_height);
 
 #ifdef __cplusplus
 }
