+++ realgui/widget/gui_text/gui_scroll_text.h
@@ -128,7 +128,7 @@ void gui_scroll_text_set(gui_scroll_text_t *_this,
                          FONT_SRC_TYPE      text_type,
                          gui_color_t        color,
                          uint16_t           length,
-                         uint8_t            font_size);
+                         uint16_t           font_size);
 
 /** * @brief set font encoding
  * @note utf-8 or unicode
